+++ src/sepsis_prediction_long_term/utils.py
@@ -1,0 +1,122 @@
+import os
+import time
+import numpy as np
+import torch
+
+
+class AverageMeter(object):
+    """Computes and stores the average and current value"""
+
+    def __init__(self):
+        self.reset()
+
+    def reset(self):
+        self.val = 0
+        self.avg = 0
+        self.sum = 0
+        self.count = 0
+
+    def update(self, val, n=1):
+        self.val = val
+        self.sum += val * n
+        self.count += n
+        self.avg = self.sum / self.count
+
+
+def compute_batch_accuracy(output, target):
+    """Computes the accuracy for a batch"""
+    with torch.no_grad():
+        batch_size = target.size(0)
+        _, pred = output.max(1)
+        correct = pred.eq(target).sum()
+
+        return correct * 100.0 / batch_size
+
+
+def train(model, device, data_loader, criterion, optimizer, epoch, print_freq=10):
+    batch_time = AverageMeter()
+    data_time = AverageMeter()
+    losses = AverageMeter()
+    accuracy = AverageMeter()
+
+    model.train()
+
+    end = time.time()
+    for i, (input, target) in enumerate(data_loader):
+        # measure data loading time
+        data_time.update(time.time() - end)
+
+        if isinstance(input, tuple):
+            input = tuple([e.to(device) if type(e) == torch.Tensor else e for e in input])
+        else:
+            input = input.to(device)
+        target = target.to(device)
+
+        optimizer.zero_grad()
+        output = model(input)
+        loss = criterion(output, target)
+        assert not np.isnan(loss.item()), 'Model diverged with loss = NaN'
+
+        loss.backward()
+        optimizer.step()
+
+        # measure elapsed time
+        batch_time.update(time.time() - end)
+        end = time.time()
+
+        losses.update(loss.item(), target.size(0))
+        accuracy.update(compute_batch_accuracy(output, target).item(), target.size(0))
+
+        if i % print_freq == 0:
+            print('Epoch: [{0}][{1}/{2}]\t'
+                  'Time {batch_time.val:.3f} ({batch_time.avg:.3f})\t'
+                  'Data {data_time.val:.3f} ({data_time.avg:.3f})\t'
+                  'Loss {loss.val:.4f} ({loss.avg:.4f})\t'
+                  'Accuracy {acc.val:.3f} ({acc.avg:.3f})'.format(
+                epoch, i, len(data_loader), batch_time=batch_time,
+                data_time=data_time, loss=losses, acc=accuracy))
+
+    return losses.avg, accuracy.avg
+
+
+def evaluate(model, device, data_loader, criterion, print_freq=10):
+    batch_time = AverageMeter()
+    losses = AverageMeter()
+    accuracy = AverageMeter()
+
+    results = []
+
+    model.eval()
+
+    with torch.no_grad():
+        end = time.time()
+        for i, (input, target) in enumerate(data_loader):
+
+            if isinstance(input, tuple):
+                input = tuple([e.to(device) if type(e) == torch.Tensor else e for e in input])
+            else:
+                input = input.to(device)
+            target = target.to(device)
+
+            output = model(input)
+            loss = criterion(output, target)
+
+            # measure elapsed time
+            batch_time.update(time.time() - end)
+            end = time.time()
+
+            losses.update(loss.item(), target.size(0))
+            accuracy.update(compute_batch_accuracy(output, target).item(), target.size(0))
+
+            y_true = target.detach().to('cpu').numpy().tolist()
+            y_pred = output.detach().to('cpu').max(1)[1].numpy().tolist()
+            results.extend(list(zip(y_true, y_pred)))
+
+            if i % print_freq == 0:
+                print('Test: [{0}/{1}]\t'
+                      'Time {batch_time.val:.3f} ({batch_time.avg:.3f})\t'
+                      'Loss {loss.val:.4f} ({loss.avg:.4f})\t'
+                      'Accuracy {acc.val:.3f} ({acc.avg:.3f})'.format(
+                    i, len(data_loader), batch_time=batch_time, loss=losses, acc=accuracy))
+
+    return losses.avg, accuracy.avg, results
